playerdata just broke completely

+++ PitMystics/src/main/java/dev/kyro/pitsim/controllers/objects/PitPlayer.java
@@ -278,7 +278,7 @@ public class PitPlayer {
 		}
 	}
 
-	public double getKills() {
+	public int getKills() {
 		return kills;
 	}
